Fixes new updates (#19)
* Security Fixes

* Modularisation

* Modularisation-Schema

* Modularisation - Validate module

* Modularisation - Base Validation

* Modularisation - Validators

* Modularisation - Final

* SafeRegex Honesty Fix

* chore: enable lint-staged pre-commit hook

* ESlint upgrade

* Logs updated

* Renaming

* Renaming
