--- .eslintrc.js
@@ -1,23 +1,0 @@
-module.exports = {
-    env: {
-        browser: true,
-        es2021: true,
-        node: true,
-        jest: true
-    },
-    extends: [
-        'eslint:recommended'
-    ],
-    parserOptions: {
-        ecmaVersion: 12,
-        sourceType: 'module'
-    },
-    rules: {
-        'indent': ['error', 2],
-        'linebreak-style': ['error', 'unix'],
-        'quotes': ['error', 'single'],
-        'semi': ['error', 'always'],
-        'no-unused-vars': 'error',
-        'no-console': 'warn'
-    }
-};
