+++ .github/workflows/ci.yml
@@ -1,4 +1,5 @@
 name: CI/CD Pipeline
+
 on:
   push:
     branches: [main, develop]
@@ -13,30 +14,53 @@ jobs:
     permissions:
       contents: read
     strategy:
+      fail-fast: false
       matrix:
-        node-version: [16.x, 18.x, 20.x]
+        node-version: [18.x, 20.x, 22.x, 24.x]
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
+
       - name: Use Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v4
         with:
           node-version: ${{ matrix.node-version }}
           cache: 'npm'
+
       - name: Install dependencies
         run: npm ci
-      - name: Run linting
-        run: npm run lint
+
       - name: Run tests
         run: npm run test:coverage
+
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v5
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          file: ./coverage/lcov.info
+          files: ./coverage/lcov.info
           flags: unittests
           name: codecov-umbrella
           fail_ci_if_error: false
+
+  lint:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Use Node.js 22
+        uses: actions/setup-node@v4
+        with:
+          node-version: 22
+          cache: 'npm'
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Run linting
+        run: npm run lint
 
   security:
     runs-on: ubuntu-latest
@@ -45,38 +69,46 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
-      - name: Use Node.js 18
+
+      - name: Use Node.js 22
         uses: actions/setup-node@v4
         with:
-          node-version: 18
+          node-version: 22
           cache: 'npm'
+
       - name: Install dependencies
         run: npm ci
+
       - name: Run security audit
         run: npm audit --audit-level=moderate
+
       - name: Check for vulnerabilities
         run: npx audit-ci --moderate
 
   publish:
-    needs: [test, security]
+    needs: [test, lint, security]
     runs-on: ubuntu-latest
     if: github.event_name == 'release' && github.event.action == 'published'
     permissions:
-      contents: write # Required for uploading release assets
-      id-token: write # Required for provenance generation
+      contents: write   # Required for uploading release assets
+      id-token: write    # Required for provenance generation
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
-      - name: Use Node.js 18
+
+      - name: Use Node.js 22
         uses: actions/setup-node@v4
         with:
-          node-version: 18
+          node-version: 22
           cache: 'npm'
           registry-url: 'https://registry.npmjs.org'
+
       - name: Install dependencies
         run: npm ci
+
       - name: Run tests
         run: npm test
+
       - name: Update package version to match release tag
         run: |
           # Extract version from release tag (removes 'v' prefix if present)
@@ -90,19 +122,27 @@ jobs:
             fs.writeFileSync('package.json', JSON.stringify(pkg, null, 2) + '\n');
             console.log('Updated package.json version to:', pkg.version);
           "
+
       - name: Check if version already exists
+        id: version_check
         run: |
           PACKAGE_NAME=$(node -p "require('./package.json').name")
           PACKAGE_VERSION=$(node -p "require('./package.json').version")
-          if npm view $PACKAGE_NAME@$PACKAGE_VERSION version 2>/dev/null; then
+          if npm view "$PACKAGE_NAME@$PACKAGE_VERSION" version >/dev/null 2>&1; then
             echo "Version $PACKAGE_VERSION already exists on npm. Skipping publish."
-            exit 0
+            echo "exists=true" >> "$GITHUB_OUTPUT"
+          else
+            echo "exists=false" >> "$GITHUB_OUTPUT"
           fi
+
       - name: Publish to NPM
+        if: steps.version_check.outputs.exists == 'false'
         run: npm publish --provenance --access public
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
+
       - name: Create and Upload Release Assets
+        if: steps.version_check.outputs.exists == 'false'
         run: |
           tar -czf snap-validate-${{ github.event.release.tag_name }}.tar.gz src/ README.md package.json LICENSE
           gh release upload ${{ github.event.release.tag_name }} snap-validate-${{ github.event.release.tag_name }}.tar.gz
