+++ src/validators/alphanumeric.js
@@ -1,0 +1,9 @@
+const { BaseValidator } = require('../core/BaseValidator');
+
+const alphanumeric = (value) => {
+  return new BaseValidator(value)
+    .required('This field is required')
+    .pattern(/^[a-zA-Z0-9]+$/, 'Only letters and numbers are allowed');
+};
+
+module.exports = { alphanumeric };
